🏗️ Move `OrganizationForm` component into new dedicated feature folder
